// frontend/src/lib/api/models/UpdatePreferencesRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdatePreferencesRequest = {
    preferredLanguage: UpdatePreferencesRequest.preferredLanguage;
};
export namespace UpdatePreferencesRequest {
    export enum preferredLanguage {
        EN = 'en',
        AR = 'ar',
        ZH = 'zh',
        ES = 'es',
        FR = 'fr',
    }
}